requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
livekit==0.17.0
livekit-api==0.5.0
pydantic==2.5.3
pydantic-settings==2.1.0
